refactor: 拆分cli以便后续拓展功能
